+++ HRLeaveManagement.Application/Features/LeaveRequests/Handlers/Commands/DeleteLeaveRequestCommandHandler.cs
@@ -1,6 +1,8 @@
 ﻿using AutoMapper;
+using HRLeaveManagement.Application.Exceptions;
 using HRLeaveManagement.Application.Features.LeaveRequests.Requests.Commands;
 using HRLeaveManagement.Application.Persistence.Contracts;
+using HRLeaveManagement.Domain;
 using MediatR;
 using System.Threading;
 using System.Threading.Tasks;
@@ -20,8 +22,10 @@ namespace HRLeaveManagement.Application.Features.LeaveRequests.Handlers.Commands
         public async Task<Unit> Handle(DeleteLeaveRequestCommand request, CancellationToken cancellationToken)
         {
             var leaveRequest = await leaveRequestRepository.Get(request.Id);
-            await leaveRequestRepository.Delete(leaveRequest);
+            if(leaveRequest == null)
+                throw new NotFoundException(nameof(LeaveRequest), request.Id);
 
+            await leaveRequestRepository.Delete(leaveRequest);
             return Unit.Value;
         }
     }
